Code restructured into a Python package, improved PEP8 compliance
- CharmPy core files have been broken into submodules and organized
  into a Python package.

- Many modules, particularly all the top-level modules in charmpy
  package now have PEP8-compliant indentation (note this doesn't mean
  that they are fully PEP8-compliant).

- Other changes have been made to several modules to improve PEP8
  compliance.
