+++ charmpy/__init__.py
@@ -1,0 +1,12 @@
+
+from .charm import charm, readonlies, Options
+Reducer = charm.reducers
+
+CkMyPe = charm.myPe
+CkNumPes = charm.numPes
+CkExit = charm.exit
+CkAbort = charm.abort
+
+from .entrymethod import when, threaded
+
+from .chare import Chare, Group, Array, ArrayMap
