+++ docs/contact.rst
@@ -5,7 +5,7 @@ Contact
 You can contact us in the `mailing list`_ for discussion of any topics related to
 CharmPy or Charm++.
 
-.. _mailing list: https://lists.cs.illinois.edu/lists/info/charm
+.. _mailing list: charm@cs.illinois.edu
 
 Please use the GitHub page to report and track issues (https://github.com/UIUC-PPL/charmpy)
